kua policy

+++ app/Http/Livewire/Kua/Form.php
@@ -24,21 +24,28 @@ class Form extends Kua
 
     public function render()
     {
+        $this->authorize('viewAny', new ModelsKua);
+
         $typologies = Typology::get(['id', 'name']);
         return view('livewire.kua.form', compact('typologies'));
     }
 
     public function create()
     {
+        $this->authorize('create', new ModelsKua);
+
         $this->openCloseModal();
     }
 
     public function storeOrUpdate()
     {
+
         $data = $this->validate();
         $data['created_by'] = auth()->user()->id;
 
-        ModelsKua::updateOrCreate(['id' => $this->kuaId], $data);
+        $kua = ModelsKua::updateOrCreate(['id' => $this->kuaId], $data);
+
+        $this->authorize('update', $kua);
 
         session()->flash('message', $this->kuaId ? 'Data KUA ' . $this->name. ' berhasil diubah' : 'Data KUA '.$this->name.' berhasil ditambahkan');
         $this->closeModal();
@@ -47,8 +54,12 @@ class Form extends Kua
 
     public function edit($id)
     {
+
         $this->kuaId        = $id;
         $kua                = ModelsKua::findOrFail($id);
+
+        $this->authorize('update', $kua);
+
         $this->name         = $kua->name;
         $this->typology_id  = $kua->typology_id;
         $this->openCloseModal();
@@ -58,6 +69,9 @@ class Form extends Kua
     {
         $this->kuaIdDelete  = $id;
         $kua                = ModelsKua::findOrFail($id);
+
+        $this->authorize('delete', $kua);
+
         $this->name         = $kua->name;
         $this->openCloseModal();
     }
@@ -65,6 +79,9 @@ class Form extends Kua
     public function destroy()
     {
         $kua = ModelsKua::findOrFail($this->kuaIdDelete);
+
+        $this->authorize('delete', $kua);
+
         $kua->delete();
         session()->flash('message', 'Data KUA ' . $kua->name .' berhasil dihapus');
         $this->closeModal();
